+++ lib/action_cable/subscription_adapter/enhanced_postgresql.rb
@@ -214,3 +214,5 @@ module ActionCable
     end
   end
 end
+
+require_relative "enhanced_postgresql/reliable_broadcasting"